+++ lib/slack/real_time/config.rb
@@ -36,15 +36,15 @@ module Slack
       private
 
       def detect_concurrency
-        %i[Eventmachine Celluloid].each do |concurrency|
+        %i[Async Eventmachine Celluloid].each do |concurrency|
           begin
             return Slack::RealTime::Concurrency.const_get(concurrency)
-          rescue LoadError, NameError
+          rescue LoadError, NameError => e
             false # could not be loaded, missing dependencies
           end
         end
 
-        raise NoConcurrencyError, 'Missing concurrency. Add faye-websocket or celluloid-io to your Gemfile.'
+        raise NoConcurrencyError, 'Missing concurrency. Add async-websocket, faye-websocket or celluloid-io to your Gemfile.'
       end
     end